Fixed #34343 -- Moved built-in templates to filesystem.
