+++ tests/view_tests/tests/test_static.py
@@ -1,6 +1,7 @@
 import mimetypes
 import unittest
 from os import path
+from unittest import mock
 from urllib.parse import quote
 
 from django.conf.urls.static import static
@@ -8,7 +9,7 @@ from django.core.exceptions import ImproperlyConfigured
 from django.http import FileResponse, HttpResponseNotModified
 from django.test import SimpleTestCase, override_settings
 from django.utils.http import http_date
-from django.views.static import was_modified_since
+from django.views.static import directory_index, was_modified_since
 
 from .. import urls
 from ..urls import media_dir
@@ -151,6 +152,18 @@ class StaticTests(SimpleTestCase):
     def test_index_custom_template(self):
         response = self.client.get("/%s/" % self.prefix)
         self.assertEqual(response.content, b"Test index")
+
+    def test_template_encoding(self):
+        """
+        The template is loaded directly, not via a template loader, and should
+        be opened as utf-8 charset as is the default specified on template
+        engines.
+        """
+        from django.views.static import Path
+
+        with mock.patch.object(Path, "open") as m:
+            directory_index(mock.MagicMock(), mock.MagicMock())
+            m.assert_called_once_with(encoding="utf-8")
 
 
 class StaticHelperTest(StaticTests):
